refactor(mesh-v2): extract network-filter mixin from mesh-service
issue #566 の責務分割 1/N。ネットワークエラー分類・i-Filter 検出・
WebSocket 可用性判定の関連メソッドを `network-filter.js` に切り出し、
ミックスイン経由で `MeshV2Service.prototype` に再アタッチする。

ミックスインパターンを採用した理由:
- すべての `this` 参照がそのまま動くため、ロジックの動作変更は一切なし
- public API (createGroup, joinGroup, testWebSocket, ...) も変更なし
- 既存テストは無修正で pass
- 各分割ファイルが独立したファイルとして閲覧・テスト可能

切り出し対象:
- shouldDisconnectOnError
- isNetworkFilterError
- _createSimulated503Error
- _checkSimulateNetworkFilter
- testWebSocket
- DISCONNECT_ERROR_TYPES 定数

mesh-service.js: 1479 → 1332 行 (-147)
network-filter.js: 173 行 (新規)
